Fix flywheel not reporting a zero goal when turned off
GetGoal() was returning the LQR reference instead of nextR, and that
reference isn't updated when nextR == 0. Now nextR is returned.

Furthermore, the LQR is now updated even when nextR == 0. This fixes a
bug where the flywheel would turn back on when the robot was disabled
then reenabled after a shooting sequence.