Refactor ControllerBase
Controllers now take their measurements in Update() instead of using a
separate setter. This was done because sensors can now be mocked to
test the controllers.

All logging and control input storage was moved into ControllerBase to
deduplicate it. Also some controllers originally used r_{k+1} for
feedback calculations instead r_k, so that was fixed.