Run controllers with TimedRobot::AddPeriodic() instead of Notifier
This fixes some controller determinism issues in simulation and unit
testing.

We had to shim x86 REV Spark Max library as well. The native library
crashes in the CANSparkMax heartbeat Notifier on GitHub Actions
(probably due to a destruction race). This shims out the x86 library
entirely to avoid it.