Revise autonomous to work better
Make the drivetrain's center start on the initiation line. This makes
robot setup on the competition field easier.

Add comments to each autonomous state describing what the state is for
and how it's reached.

Fix six-ball and Intake autonomous modes so they are more functional,
and make sure they start and reach where they should.

Closes #91, Closes #92