Refactor autonomous verifications
All verifications at the end of autonomous except the state machine
check were moved into a separate member function and called in the
autonomous test. The state machine check now occurs at 14.97s instead of
14.5s.

This process uncovered that the no-op autonomous was moving the turret.
Additional tests were added for that and the turret is now put in manual
mode for the no-op autonomous.