Replace autonomous mode state machines with actor model
This make the autonomous mode logic procedural instead of imperative,
which is much easier to read and write. AutonomousChooser uses condition
variables to ensure the main robot thread and separate autonomous mode
thread don't run concurrently to avoid potential race conditions and
avoid the need for synchronization in user code.