Make turret to climber transition safer (#169)
Tested the climber which moved the turret out of the way correctly.
After attempting to move the climber again the turret when past both its
CCW limit and the magnetic limit switch decapitating itself. Setting the
position to pi / 2 radians should lead to less chance of a decapitation
and more time to stop it if it were to happen.

Originally the turret went to pi / 2 radians because the CCW limit was
set there. After the CCW limit was moved to 3 pi / 2 radians because the
turret was able to safely move farther. Now set the turret to move to pi
/ 2 radians since it's not needed to move further than that.