Fix turret configuration space violation in collision avoidance
We reexpanded the configuration space when the climber was below 1.5 in
even though the climber was being moved up.

The collision avoidance test was moved into RobotTest so it's using the
same robot environment as other integration tests. The test was
rewritten to check turret positions instead of just motor voltages.